% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{add_inputdata}
\alias{add_inputdata}
\title{Add user-input data to Kotka upload}
\usage{
add_inputdata(k, x)
}
\arguments{
\item{k}{A data frame containing the Kotka upload.}

\item{x}{A data frame containing the user input.}
}
\value{
The data frame containing the Kotka upload (\code{k}) with user input added to the matching columns.
}
\description{
This is a helper function used by \code{\link[=make_upload]{make_upload()}}. Adds data columns input by the user to a Kotka upload. Ignores column "sample", tries to find the right Kotka column for everything else.
}
\details{
Any columns that can't be matched to a Kotka column are ignored.
}
\seealso{
\code{\link[=make_upload]{make_upload()}}
}
