% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{add_malaisedata}
\alias{add_malaisedata}
\title{Add Peruvian and Ugandan data to Kotka upload}
\usage{
add_malaisedata(k, x)
}
\arguments{
\item{k}{A data frame containing the Kotka upload.}

\item{x}{A data frame returned by \code{\link[=get_labeldata]{get_labeldata()}} or \code{\link[=verify_labeldata]{verify_labeldata()}}.}
}
\value{
The data frame containing the Kotka upload (\code{k}) with Peruvian and Ugandan data added.
}
\description{
This is a helper function used by \code{\link[=make_kotkaupload]{make_kotkaupload()}}. Add data on Peruvian and Ugandan wasps caught by Malaise trapping to a Kotka upload. Add basic data from a template file, and sample data by matching each wasp's sample ID with the sample it came from.
}
\seealso{
\code{\link[=make_kotkaupload]{make_kotkaupload()}}, and \code{\link[=read]{read()}} which is used by this function
}
