% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/on_load_package.R
\name{.onLoad}
\alias{.onLoad}
\title{Load malaise sample data when the package is loaded}
\usage{
.onLoad(libname, pkgname)
}
\arguments{
\item{libname}{Not used, there only to make sure this function is called properly when the package is loaded.}

\item{pkgname}{Not used, there only to make sure this function is called properly when the package is loaded.}
}
\description{
Loads the Malaise sample data so that the package's code and the user can see it. The sample data is in two csv files in the package's external data, reads both of them and saves as 'm' (list of samples with basic data) and 'm_kotka' (complete sample data in Kotka format)
}
