% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_package_and_data.R
\docType{data}
\name{example_labels}
\alias{example_labels}
\title{Example labels of wasps from Peru and Ecuador}
\format{
A data frame with 188 rows (one row for each wasp) and 1 variables:
\describe{
\item{label}{Text of labels in the format expected by Kotka. Each label on a separate row, separated by carriage returns.}
}
}
\usage{
example_labels
}
\description{
Labels of Amazonian wasps collected in Peru and Ecuador. These give a good example of what kind of labels there are: Ecuadorian canopy fogging labels, Peruvian labels from 1998-2011, quite a few mistypes, mistakes etc.. These are real labels of Banchinae wasps.
}
\keyword{datasets}
