% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_d_begin}
\alias{get_d_begin}
\title{Get the start date from Peruvian wasp labels}
\usage{
get_d_begin(x, d_end)
}
\arguments{
\item{x}{A string giving the collecting dates of a Peruvian label (e.g. \emph{1.-15.12.2000}).}

\item{d_end}{An object of class \code{Date} giving the end date (e.g. \emph{2000-12-15}). This is used to get the start month and year if they are not written out in full.}
}
\value{
An object of class \code{Date} giving the start date.
}
\description{
This is a helper function used by \code{\link[=get_dates]{get_dates()}}. Figure out what the start date is for Peruvian labels, whose date is typically written along the lines of \emph{1.-15.12.2000}.
}
\seealso{
\code{\link[=get_dates]{get_dates()}}
}
