% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_dates}
\alias{get_dates}
\title{Get dates from Peruvian or Ecuadorian wasp labels}
\usage{
get_dates(lab)
}
\arguments{
\item{lab}{A character vector of label texts.}
}
\value{
A data frame with two columns, giving the start and end dates as objects of class \code{Date}.
}
\description{
This is a helper function used by \code{\link[=get_labeldata]{get_labeldata()}}. Extract those strings in Peruvian and Ecuadorian labels which represent dates (e.g. \emph{1.-15.12.2000} or \emph{22. Oct 1998}).
}
\seealso{
\code{\link[=get_labeldata]{get_labeldata()}}, and \code{\link[=getreg]{getreg()}} and \code{\link[=get_d_begin]{get_d_begin()}} which are used by this function
}
