% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{getreg}
\alias{getreg}
\title{Get strings which match a regular expression}
\usage{
getreg(reg, x)
}
\arguments{
\item{reg}{A string containing a regular expression.}

\item{x}{A string or character vector to search with the regular expression.}
}
\value{
A string or character vector giving the string(s) that matched the regular expression.
}
\description{
This is a helper function used by \code{\link[=get_dates]{get_dates()}} to extract dates from labels. Look for matches to a regular expression in text, return the strings that match.
}
\seealso{
\code{\link[=get_dates]{get_dates()}}, and \code{\link[=regexpr]{regexpr()}} which this is largely a wrapper for
}
