% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_package_and_data.R
\docType{data}
\name{m}
\alias{m}
\title{Malaise samples collected in Peru, Uganda and elsewhere (overview)}
\format{
A data frame with 1540 rows (one row for each sample) and 12 variables:
\describe{
\item{sample}{Sample ID, plaintext form used at the Turku Zoological Museum}
\item{event}{Malaise trapping that collected the sample, e.g. "Amazon 1998"}
\item{trap}{Malaise trap that collected the sample}
\item{sample_number}{Latter half of the sample ID, telling which of the trap's samples this is}
\item{date_begin}{When the Malaise sample started to be collected. Finnish date format \emph{day.month.year}.}
\item{date_end}{When the Malaise sample stopped being collected. Finnish date format \emph{day.month.year}.}
\item{MYDatasetID[0]}{Dataset identifier used by the Kotka. This dataset gives the collecting event.}
\item{MYDatasetID[1]}{Dataset identifier used by the Kotka. This dataset gives the forest type.}
\item{MYDatasetID[2]}{Dataset identifier used by the Kotka. This dataset gives the trap.}
\item{MYDatasetID[3]}{Dataset identifier used by the Kotka. This dataset marks any damaged samples.}
\item{sample_1998_2000}{How the sample ID is typically written on labels related to the Amazon 1998 and Amazon 2000 Malaise trapping.}
\item{trap_2008_2011}{How the trap is typically written on labels related to the Amazon 2008 and Amazon 2011 Malaise trapping.}
}
}
\usage{
m
}
\description{
An abbreviated version of the \code{m_kotka} dataset. Contains basic data on what Malaise samples were collected in e.g. Peru and Uganda, including how the samples might be written on labels. Used by the script to e.g. recognise sample IDs from labels.
}
\keyword{datasets}
