% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_kotkaupload.R
\name{make_kotkaupload}
\alias{make_kotkaupload}
\title{Create a file for uploading wasp data to the Kotka database}
\usage{
make_kotkaupload(x, write.csv = TRUE)
}
\arguments{
\item{x}{A data frame returned by \code{\link[=get_labeldata]{get_labeldata()}} or \code{\link[=verify_labeldata]{verify_labeldata()}}.}

\item{write.csv}{If TRUE (the default) the Kotka upload will be written to file "kotka_upload.csv", as well as returned invisibly. If FALSE, the Kotka upload will only be returned invisibly.}
}
\value{
A data frame with the Kotka upload. Also contains the columns of the input data frame (\code{x}), which the upload is based on. \cr

The columns of \code{x} are there for reference before uploading. Typically, column \code{problems} will be checked so any issues can be fixed. These columns should be deleted before uploading to Kotka, they are marked with "delete this column" and are at the start of the file.
}
\description{
Create a Kotka upload file out of the data extracted from wasp labels (or obtained from sample IDs). Handles wasps from the Malaise trapping in Peru (1998-2011), and from the canopy fogging in Ecuador. Other specimens will be included in the upload, but will only have minimal data.
}
\examples{
lab = c("cct1-141022", "PERU 1.-15.12.2000 I1/17", "Not a wasp label at all")
x = get_labeldata(lab)
x = verify_labeldata(x)
upload = make_kotkaupload(x, write.csv=FALSE)
}
\seealso{
\code{\link[=get_labeldata]{get_labeldata()}} which extracts the data used by this function, \code{\link[=verify_labeldata]{verify_labeldata()}} which checks the extracted data for problems.
}
