% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{malaise_samples}
\alias{malaise_samples}
\title{Malaise samples collected in Peru and Uganda (overview)}
\format{
A data frame with 1519 rows (one row for each sample) and 11 variables:
\describe{
\item{sample}{Sample ID, plaintext form used at the Turku Zoological Museum}
\item{event}{Malaise trapping that collected the sample, e.g. "Amazon 1998"}
\item{trap}{Malaise trap that collected the sample}
\item{sample_number}{Latter half of the sample ID, telling which of the trap's samples this is}
\item{date_begin}{When the Malaise sample started to be collected. Finnish date format \emph{day.month.year}.}
\item{date_end}{When the Malaise sample stopped being collected. Finnish date format \emph{day.month.year}.}
\item{sample_1998_2000}{How the sample ID is typically written on labels related to the Amazon 1998 and Amazon 2000 Malaise trapping.}
\item{MYDatasetID[0]}{Dataset identifier used by the Kotka. This dataset gives the collecting event.}
\item{MYDatasetID[1]}{Dataset identifier used by the Kotka. This dataset gives the forest type.}
\item{MYDatasetID[2]}{Dataset identifier used by the Kotka. This dataset gives the trap.}
\item{MYDatasetID[3]}{Dataset identifier used by the Kotka. This dataset marks any damaged samples.}
}
}
\usage{
malaise_samples
}
\description{
An abbreviated version of the \code{malaise_samples_kotka_format} dataset. Contains basic data on what Malaise samples were collected in Peru and Uganda, including how the samples might be written on labels. Used by the script to e.g. recognise sample IDs from labels.
}
\keyword{datasets}
