% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{malaise_samples_kotka_format}
\alias{malaise_samples_kotka_format}
\title{Malaise samples collected in Peru and Uganda (Kotka format)}
\format{
A data frame with 1519 rows (one row for each sample) and 46 variables. The variables are described in \url{https://kotka.luomus.fi/documentation/field}.
}
\usage{
malaise_samples_kotka_format
}
\description{
Data on the Malaise samples collected in Peru and Uganda, in the format Kotka uses. Contains almost everything that is in Kotka (some irrelevant columns may have been omitted). Used by the script to get the collecting dates, coordinates etc data for wasps caught in these samples.
}
\details{
This file has been downloaded from the Kotka Collection Management System in Excel format. If a new version is ever needed:
\itemize{
\item download from Kotka
\item remove second header row
\item convert to csv
\item read in to R keeping the column names intact, e.g. \code{read.csv(file, colClasses="character", check.names=F)}
}
}
\keyword{datasets}
