% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{read}
\alias{read}
\title{Read csv files that have multiple headers}
\usage{
read(file, nheaders = 2, ...)
}
\arguments{
\item{file}{The name of the file to be read from.}

\item{nheaders}{An integer giving the number of header rows in the file. Only the first header row will be kept.}

\item{...}{Further arguments to be passed to \code{\link[utils:read.table]{utils::read.csv()}}.}
}
\value{
A data frame with the contents of the file.
}
\description{
This is a helper function used by \code{\link[=add_malaisedata]{add_malaisedata()}}. Help read the kind of csv files typically used by Kotka, which contain more than one header row (and are otherwise hard to read with \code{\link[utils:read.table]{utils::read.csv()}}). Any extra header rows are discarded.
}
\seealso{
\code{\link[=add_malaisedata]{add_malaisedata()}}, and \code{\link[utils:read.table]{utils::read.csv()}} which this is a wrapper for
}
