% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{verify_data}
\alias{verify_data}
\title{Verify specimen data}
\usage{
verify_data(x)
}
\arguments{
\item{x}{A data frame, e.g. that created by \code{\link[=get_labeldata]{get_labeldata()}} or received by \code{\link[=make_upload]{make_upload()}}. Must include column "sample".}
}
\value{
The input data frame with extra column "sample_problem".
}
\description{
This is a helper function used by \code{\link[=make_upload]{make_upload()}} and \code{\link[=get_labeldata]{get_labeldata()}}. Checks that the data they are handling looks right, and highlights any problems.
}
\details{
Any problems, such as samples not existing or sample dates not matching dates on the label, will be added to a separate column to the data.
}
\seealso{
\code{\link[=get_labeldata]{get_labeldata()}} whose data this function checks, \code{\link[=make_upload]{make_upload()}} which creates a Kotka upload file out of the returned data frame.
}
