% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_labeldata.R
\name{verify_labeldata}
\alias{verify_labeldata}
\title{Verify data extracted from wasp labels}
\usage{
verify_labeldata(x)
}
\arguments{
\item{x}{A data frame returned by \code{\link[=get_labeldata]{get_labeldata()}}.}
}
\value{
The input data frame with an extra column \code{problems}.
}
\description{
Check that the data extracted by \code{\link[=get_labeldata]{get_labeldata()}} looks right, and highlight any problems.
}
\details{
Typically used before feeding in the data to function \code{\link[=make_kotkaupload]{make_kotkaupload()}}. Any problems will be added as a separate column to the data. When the Kotka upload file has been made, it is easy to skim through the problem column and fix any issues before uploading.
Checks include:
\itemize{
\item That a start date was extracted and the dates match the sample the wasp came from.
\item That the wasp came from a recognised Peruvian or Ugandan Malaise sample, or from Ecuadorian canopy fogging. (If not, most data will be blank.)
\item That the sex of the wasp (male/female) was detected.
}
}
\examples{
\dontrun{
lab = c("cct1-141022", "PERU 1.-15.12.2000 I1/17")
x = get_labeldata(lab)
x = verify_labeldata(x)
}
}
\seealso{
\code{\link[=get_labeldata]{get_labeldata()}} whose data this function checks, \code{\link[=make_kotkaupload]{make_kotkaupload()}} which creates a Kotka upload file out of the returned data frame.
}
