% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{add}
\alias{add}
\title{Add any problems to the problem list}
\usage{
add(problems, add = "")
}
\arguments{
\item{problems}{A list of character vectors, with one list item per label. Each vector item is a separate problem.}

\item{add}{The string to be added to the each vector in the list. This will be some kind of message detailing the problem.}
}
\value{
The input list (\code{problems}) with the message added to each list item.
}
\description{
This is a helper function used by \code{\link[=verify_labeldata]{verify_labeldata()}}. Add any problem messages of a wasp label to the list of problems.
}
\seealso{
\code{\link[=verify_labeldata]{verify_labeldata()}}
}
