% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{add_ecuadordata}
\alias{add_ecuadordata}
\title{Add Ecuadorian data to Kotka upload}
\usage{
add_ecuadordata(k, x)
}
\arguments{
\item{k}{A data frame containing the Kotka upload.}

\item{x}{A data frame returned by \code{\link[=get_labeldata]{get_labeldata()}} or \code{\link[=verify_labeldata]{verify_labeldata()}}.}
}
\value{
The data frame containing the Kotka upload (\code{k}) with Ecuadorian data added.
}
\description{
This is a helper function used by \code{\link[=make_kotkaupload]{make_kotkaupload()}}. Add basic Ecuadorian data to a Kotka upload.
}
\seealso{
\code{\link[=make_kotkaupload]{make_kotkaupload()}}
}
