% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{add_malaisedata}
\alias{add_malaisedata}
\title{Add Peruvian and Ugandan data to Kotka upload}
\usage{
add_malaisedata(k, x, subsample = FALSE)
}
\arguments{
\item{k}{A data frame containing the Kotka upload.}

\item{x}{A data frame containing the user input. If there is no column "sample", this function does nothing.}

\item{subsample}{If TRUE, gives data for subsamples instead of for individual pinned insects.}
}
\value{
The data frame containing the Kotka upload (\code{k}) with Peruvian and Ugandan data added.
}
\description{
This is a helper function used by \code{\link[=make_upload]{make_upload()}}. Adds data on Peruvian and Ugandan wasps caught by Malaise trapping to a Kotka upload. Adds basic data from a template file, and sample data by matching each wasp's sample ID with the sample it came from.
}
\seealso{
\code{\link[=make_upload]{make_upload()}}, and \code{\link[=read]{read()}} which is used by this function
}
