% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_labeldata.R
\name{get_labeldata}
\alias{get_labeldata}
\title{Extract data from wasp labels}
\usage{
get_labeldata(lab)
}
\arguments{
\item{lab}{A character vector of label texts.}
}
\value{
A data frame with the labels and corresponding data.
}
\description{
Extract dates, sample ID and other data from the wasp labels typically found at the museum. Handles wasp labels from the Malaise trapping in Peru (1998-2011), and from the canopy fogging in Ecuador. The resulting data frame can be used as is, or used as the basis for creating a Kotka upload.
}
\examples{
\dontrun{
lab = c(  
"cct1-141022",  
"h1/1",  
"PERU, Allpahuayo 1.-15.12.2000, Sääksjärvi I.E I1/17",  
"ECUADOR, Tiputini, 22. Oct 1998, Canopy fogging Lot# 1966 Meniscomorpha sp. 2"  
)
x = get_labeldata(lab)
}
}
\seealso{
\code{\link[=verify_labeldata]{verify_labeldata()}} which checks the results, \code{\link[=make_kotkaupload]{make_kotkaupload()}} which creates a Kotka upload file out of the returned (and optionally verified) data frame.
}
